perf: delay version footer work
